requirements.txt
flask
flask-cors
torch
pillow
numpy
opencv-python
